requirements.txt


requests==2.22.0
PyExecJS==1.5.1
